hide titter by nonarticle
